Changed popup button style and functionality
